adaptation moodle3.3 avec mise à jour d'editpdf v201705

+++ classes/task/convert_submissions.php
@@ -1,4 +1,5 @@
 <?php
+
 // This file is part of Moodle - http://moodle.org/
 //
 // Moodle is free software: you can redistribute it and/or modify
@@ -21,10 +22,12 @@
  * @copyright  2016 Damyon Wiese
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
 namespace assignfeedback_editpdfplus\task;
 
 use core\task\scheduled_task;
 use assignfeedback_editpdfplus\document_services;
+use assignfeedback_editpdfplus\combined_document;
 use context_module;
 use assign;
 
@@ -91,18 +94,44 @@ class convert_submissions extends scheduled_task {
             }
 
             mtrace('Convert ' . count($users) . ' submission attempt(s) for assignment ' . $assignmentid);
+            $keepinqueue = false;
             foreach ($users as $userid) {
-                document_services::get_page_images_for_attempt($assignment,
-                                                               $userid,
-                                                               $attemptnumber,
-                                                               true);
-                document_services::get_page_images_for_attempt($assignment,
-                                                               $userid,
-                                                               $attemptnumber,
-                                                               false);
+                /* document_services::get_page_images_for_attempt($assignment,
+                  $userid,
+                  $attemptnumber,
+                  true);
+                  document_services::get_page_images_for_attempt($assignment,
+                  $userid,
+                  $attemptnumber,
+                  false); */
+                $combineddocument = document_services::get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
+                $status = $combineddocument->get_status();
+
+                switch ($status) {
+                    case combined_document::STATUS_READY:
+                    case combined_document::STATUS_PENDING_INPUT:
+                        // The document has not been converted yet or is somehow still ready.
+                        $keepinqueue = true;
+                        continue;
+                }
+
+                try {
+                    document_services::get_page_images_for_attempt(
+                            $assignment, $userid, $attemptnumber, false
+                    );
+                    document_services::get_page_images_for_attempt(
+                            $assignment, $userid, $attemptnumber, true
+                    );
+                } catch (\moodle_exception $e) {
+                    mtrace('Conversion failed with error:' . $e->errorcode);
+                    $keepinqueue = true;
+                }
             }
 
-            $DB->delete_records('assignfeedback_editpp_queue', array('id' => $record->id));
+            if (!$keepinqueue) {
+                // Remove from queue unless requested not to.
+                $DB->delete_records('assignfeedback_editpp_queue', array('id' => $record->id));
+            }
         }
     }
